Refactor of watchdogs
Refactos watchdogs to match that of the autoMakeMKV paradigm where
a new 'watchdogs' subpackage is created with a BaseWatchdog class
in the base.py module that is then subclassed by the Watchdog
classin the linux.py module. This should make it a little easier
to create a windows.py module for other operating system
