+++ src/autoripper/watchdogs/__init__.py
@@ -1,0 +1,18 @@
+"""
+Utilities for ripping titles
+
+"""
+
+import signal
+from threading import Event
+
+KEY = 'DEVNAME'
+CHANGE = 'DISK_MEDIA_CHANGE'
+STATUS = "ID_CDROM_MEDIA_STATE"
+EJECT = "DISK_EJECT_REQUEST"  # This appears when initial eject requested
+READY = "SYSTEMD_READY"  # This appears when disc tray is out
+
+RUNNING = Event()
+
+signal.signal(signal.SIGINT, lambda *args: RUNNING.set())
+signal.signal(signal.SIGTERM, lambda *args: RUNNING.set())
